+++ packages/sv/src/addons/paraglide.ts
@@ -94,7 +94,7 @@ export default defineAddon({
 			`src/hooks.${language}`,
 			transforms.script(({ ast, comments, js }) => {
 				js.imports.addNamed(ast, {
-					from: `${lib}/paraglide/runtime`,
+					from: `${lib}/paraglide/runtime.js`,
 					imports: ['deLocalizeUrl']
 				});
 
@@ -117,13 +117,13 @@ export default defineAddon({
 
 				if (language === 'ts') {
 					js.imports.addNamed(ast, {
-						from: '@sveltejs/kit',
+						from: '@sveltejs/kit/hooks',
 						imports: ['Reroute'],
 						isType: true
 					});
 				} else {
 					js.common.addJsDocTypeComment(existingExport, comments, {
-						type: "import('@sveltejs/kit').Reroute"
+						type: "import('@sveltejs/kit/hooks').Reroute"
 					});
 				}
 			})
@@ -134,11 +134,11 @@ export default defineAddon({
 			`src/hooks.server.${language}`,
 			transforms.script(({ ast, comments, js }) => {
 				js.imports.addNamed(ast, {
-					from: `${lib}/paraglide/server`,
+					from: `${lib}/paraglide/server.js`,
 					imports: ['paraglideMiddleware']
 				});
 				js.imports.addNamed(ast, {
-					from: `${lib}/paraglide/runtime`,
+					from: `${lib}/paraglide/runtime.js`,
 					imports: ['getTextDirection']
 				});
 
@@ -209,7 +209,7 @@ export default defineAddon({
 			transforms.svelteScript({ language }, ({ ast, svelte, js }) => {
 				js.imports.addNamed(ast.instance.content, {
 					imports: ['locales', 'localizeHref'],
-					from: `${lib}/paraglide/runtime`
+					from: `${lib}/paraglide/runtime.js`
 				});
 				js.imports.addNamed(ast.instance.content, { imports: ['page'], from: '$app/state' });
 				js.imports.addNamed(ast.instance.content, { imports: ['resolve'], from: '$app/paths' });
@@ -248,7 +248,7 @@ export default defineAddon({
 						imports: {
 							setLocale: 'setLocale'
 						},
-						from: `${lib}/paraglide/runtime`
+						from: `${lib}/paraglide/runtime.js`
 					});
 
 					// add localized message